refactor: move RescueDetail types to dedicated feature
