+++ apps/frontend/portal/src/features/RescueDetail/api/rescueDetail.api.ts
