+++ apps/frontend/portal/src/features/RescueDetail/api/rescueDetail.queries.ts
