+++ apps/frontend/portal/src/features/RescueDetail/components/AnimalBasicInfo.tsx
