+++ apps/frontend/portal/src/features/RescueDetail/index.tsx
@@ -2,9 +2,9 @@ import { useParams } from 'react-router-dom';
 
 import { useFetchRescueDetail } from './api/rescueDetail.queries';
 import { AnimalBasicInfo } from './components/AnimalBasicInfo';
-import { RescueUpdateTimeline } from './components/RescueUpdateTimeline';
+import { RescueTimeline } from './components/RescueTimeline';
 
-const RescueDetail = () => {
+export const RescueDetail = () => {
   const { animalID } = useParams<{ animalID: string }>();
   const { data: animal } = useFetchRescueDetail(animalID ?? '');
 
@@ -12,9 +12,7 @@ const RescueDetail = () => {
     <div className="mx-auto max-w-5xl px-4 py-4 sm:px-6 md:py-6 lg:py-8">
       {animal && <AnimalBasicInfo animal={animal} />}
 
-      {animal?.updates && <RescueUpdateTimeline updates={animal.updates} />}
+      {animal?.updates && <RescueTimeline updates={animal.updates} />}
     </div>
   );
 };
-
-export { RescueDetail };
