+++ apps/frontend/portal/src/features/RescueDetail/types.ts
@@ -1,0 +1,28 @@
+import type { AnimalRescueStatus } from '@/types/AnimalType';
+
+export interface RescueUpdate {
+  id: string;
+  timestamp: string;
+  status: AnimalRescueStatus;
+  operator: {
+    id: string;
+    name: string;
+    avatar: string;
+    role: 'reporter' | 'rescuer' | 'admin';
+  };
+  content: string;
+  images?: string[];
+  location?: {
+    address: string;
+    latitude: number;
+    longitude: number;
+  };
+}
+
+export interface RescueParticipantType {
+  id: string;
+  name: string;
+  avatar?: string;
+  role: 'reporter' | 'rescuer' | 'admin';
+  joinedAt: string;
+}
